+++ android/src/main/kotlin/com/example/sentiance_flutter/Cache.java
@@ -1,5 +1,6 @@
 package com.example.sentiance_flutter;
 
+import org.json.JSONObject;
 import android.content.Context;
 import android.content.SharedPreferences;
 import androidx.annotation.Nullable;
@@ -14,8 +15,13 @@ class Cache {
     private static final String KEY_MOBILE_HEALTH_URL = "mobile_health_url";
     private static final String KEY_CRASH_DETECTION_URL = "crash_detection_url";
     private static final String KEY_USER_INSTALL_ID = "user_install_id";
+    private static final String KEY_USER_MOBILE_HEALTH_DATA ="mobile_health_data";
+    private static final String KEY_CUSTOMER_ID = "customer_id";
+    private static final String KEY_USER_ID ="user_id";
 
     private Context mContext;
+
+    
 
     Cache(Context context) {
         mContext = context;
@@ -107,6 +113,39 @@ class Cache {
     void setInitialize(String initialize) {
         SharedPreferences sharedPreferences = mContext.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
         sharedPreferences.edit().putString(KEY_INITIALIZE, initialize).apply();
+    }
+
+    void setMobileHealthData(JSONObject data)
+    {
+        SharedPreferences sharedPreferences = mContext.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
+        sharedPreferences.edit().putString(KEY_USER_MOBILE_HEALTH_DATA,data.toString()).apply();
+    }
+
+    String getMobileHealthData()
+    {
+        SharedPreferences sharedPreferences = mContext.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
+        return sharedPreferences.getString(KEY_USER_MOBILE_HEALTH_DATA, null);
+    }
+    
+    void setCustomerId(String customerId)
+    {
+        SharedPreferences sharedPreferences = mContext.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
+        sharedPreferences.edit().putString(KEY_CUSTOMER_ID, customerId).apply(); 
+    }
+    String getCustomerId()
+    {
+        SharedPreferences sharedPreferences = mContext.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
+        return sharedPreferences.getString(KEY_CUSTOMER_ID, null);
+    }
+    void setUserId(String customerId)
+    {
+        SharedPreferences sharedPreferences = mContext.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
+        sharedPreferences.edit().putString(KEY_CUSTOMER_ID, customerId).apply(); 
+    }
+    String getUserId()
+    {
+        SharedPreferences sharedPreferences = mContext.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
+        return sharedPreferences.getString(KEY_CUSTOMER_ID, null);
     }
 }